+++ src/settings/MappingProfiles/detailsSections/edit/InvoiceDetailSection/VendorInformation.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 import { Field } from 'redux-form';
 
@@ -9,10 +10,22 @@ import {
   TextField,
 } from '@folio/stripes/components';
 
+import {
+  FieldOrganization,
+  AcceptedValuesField,
+} from '../../../../../components';
+
 import { getFieldName } from '../../utils';
+import { okapiShape } from '../../../../../utils';
 import { TRANSLATION_ID_PREFIX } from '../../constants';
 
-export const VendorInformation = () => {
+export const VendorInformation = ({
+  setReferenceTables,
+  filledVendorId,
+  accountingCodeOptions,
+  onSelectVendor,
+  okapi,
+}) => {
   return (
     <Accordion
       id="vendor-information"
@@ -27,12 +40,41 @@ export const VendorInformation = () => {
           />
         </Col>
         <Col xs={4}>
-          <FormattedMessage id={`${TRANSLATION_ID_PREFIX}.invoice.vendorInformation.field.vendorName`} />
+          <FieldOrganization
+            id={filledVendorId}
+            setReferenceTables={setReferenceTables}
+            name={getFieldName(17)}
+            onSelect={onSelectVendor}
+            label={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.invoice.vendorInformation.field.vendorName`} />}
+          />
         </Col>
         <Col xs={4}>
-          <FormattedMessage id={`${TRANSLATION_ID_PREFIX}.invoice.vendorInformation.field.accountingCode`} />
+          <AcceptedValuesField
+            component={TextField}
+            name={getFieldName(18)}
+            label={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.invoice.vendorInformation.field.accountingCode`} />}
+            optionValue="value"
+            optionLabel="label"
+            wrapperLabel={`${TRANSLATION_ID_PREFIX}.wrapper.acceptedValues`}
+            acceptedValuesList={accountingCodeOptions}
+            disabled={!accountingCodeOptions.length}
+            okapi={okapi}
+          />
         </Col>
       </Row>
     </Accordion>
   );
+};
+
+VendorInformation.propTypes = {
+  okapi: okapiShape.isRequired,
+  setReferenceTables: PropTypes.func.isRequired,
+  onSelectVendor: PropTypes.func.isRequired,
+  filledVendorId: PropTypes.string,
+  accountingCodeOptions: PropTypes.arrayOf(PropTypes.object),
+};
+
+VendorInformation.defaultProps = {
+  filledVendorId: '',
+  accountingCodeOptions: [],
 };